+++ app/src/main/java/com/neuralbit/letsnote/ui/adapters/LabelRVAdapter.kt
@@ -1,4 +1,4 @@
-package com.neuralbit.letsnote.adapters
+package com.neuralbit.letsnote.ui.adapters
 
 import android.content.Context
 import android.view.LayoutInflater
