+++ app/src/main/java/com/neuralbit/letsnote/ui/addEditNote/NoteViewModel.kt
@@ -1,4 +1,4 @@
-package com.neuralbit.letsnote
+package com.neuralbit.letsnote.ui.addEditNote
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
