+++ app/src/main/java/com/neuralbit/letsnote/ui/archived/ArchivedFragment.kt
@@ -20,11 +20,11 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import com.google.gson.Gson
-import com.neuralbit.letsnote.AddEditNoteActivity
-import com.neuralbit.letsnote.Fingerprint
+import com.neuralbit.letsnote.ui.addEditNote.AddEditNoteActivity
+import com.neuralbit.letsnote.ui.addEditNote.Fingerprint
 import com.neuralbit.letsnote.Services.DeleteReceiver
-import com.neuralbit.letsnote.adapters.NoteFireClick
-import com.neuralbit.letsnote.adapters.NoteRVAdapter
+import com.neuralbit.letsnote.ui.adapters.NoteFireClick
+import com.neuralbit.letsnote.ui.adapters.NoteRVAdapter
 import com.neuralbit.letsnote.databinding.FragmentArchivedNotesBinding
 import com.neuralbit.letsnote.entities.NoteFire
 import com.neuralbit.letsnote.ui.allNotes.AllNotesViewModel
