+++ app/src/main/java/com/neuralbit/letsnote/ui/label/LabelNotesActivity.kt
@@ -1,4 +1,4 @@
-package com.neuralbit.letsnote
+package com.neuralbit.letsnote.ui.label
 
 import android.app.AlarmManager
 import android.app.AlertDialog
@@ -22,10 +22,13 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import com.google.gson.Gson
+import com.neuralbit.letsnote.ui.addEditNote.Fingerprint
+import com.neuralbit.letsnote.R
 import com.neuralbit.letsnote.Services.DeleteReceiver
-import com.neuralbit.letsnote.adapters.NoteFireClick
-import com.neuralbit.letsnote.adapters.NoteRVAdapter
+import com.neuralbit.letsnote.ui.adapters.NoteFireClick
+import com.neuralbit.letsnote.ui.adapters.NoteRVAdapter
 import com.neuralbit.letsnote.entities.NoteFire
+import com.neuralbit.letsnote.ui.addEditNote.AddEditNoteActivity
 import com.neuralbit.letsnote.ui.allNotes.AllNotesViewModel
 import com.neuralbit.letsnote.utilities.AlertReceiver
 import java.util.*
@@ -135,7 +138,7 @@ class LabelNotesActivity : AppCompatActivity() , NoteFireClick {
         }
 
         override fun onActionItemClicked(mode: ActionMode?, item: MenuItem?): Boolean {
-            if (item?.itemId == R.id.archive ){
+            if (item?.itemId == R.id.archive){
                 archiveNotes()
                 mode?.finish()
                 return true
