+++ app/src/main/java/com/neuralbit/letsnote/ui/tag/TagNotesActivity.kt
@@ -1,4 +1,4 @@
-package com.neuralbit.letsnote
+package com.neuralbit.letsnote.ui.tag
 
 import android.app.AlarmManager
 import android.app.AlertDialog
@@ -19,10 +19,13 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import com.google.gson.Gson
+import com.neuralbit.letsnote.ui.addEditNote.Fingerprint
+import com.neuralbit.letsnote.R
 import com.neuralbit.letsnote.Services.DeleteReceiver
-import com.neuralbit.letsnote.adapters.NoteFireClick
-import com.neuralbit.letsnote.adapters.NoteRVAdapter
+import com.neuralbit.letsnote.ui.adapters.NoteFireClick
+import com.neuralbit.letsnote.ui.adapters.NoteRVAdapter
 import com.neuralbit.letsnote.entities.NoteFire
+import com.neuralbit.letsnote.ui.addEditNote.AddEditNoteActivity
 import com.neuralbit.letsnote.ui.allNotes.AllNotesViewModel
 import com.neuralbit.letsnote.utilities.AlertReceiver
 import java.util.*
@@ -149,7 +152,7 @@ class TagNotesActivity : AppCompatActivity() , NoteFireClick {
         }
 
         override fun onActionItemClicked(mode: ActionMode?, item: MenuItem?): Boolean {
-            if (item?.itemId == R.id.archive ){
+            if (item?.itemId == R.id.archive){
                 archiveNotes()
                 mode?.finish()
                 return true
