+++ app/src/main/java/com/neuralbit/letsnote/ui/todos/AllTodoFragment.kt
@@ -25,12 +25,12 @@ import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.gson.Gson
-import com.neuralbit.letsnote.AddEditNoteActivity
-import com.neuralbit.letsnote.Fingerprint
+import com.neuralbit.letsnote.ui.addEditNote.AddEditNoteActivity
+import com.neuralbit.letsnote.ui.addEditNote.Fingerprint
 import com.neuralbit.letsnote.R
 import com.neuralbit.letsnote.Services.DeleteReceiver
-import com.neuralbit.letsnote.adapters.NoteFireClick
-import com.neuralbit.letsnote.adapters.NoteRVAdapter
+import com.neuralbit.letsnote.ui.adapters.NoteFireClick
+import com.neuralbit.letsnote.ui.adapters.NoteRVAdapter
 import com.neuralbit.letsnote.databinding.FragmentAllTodosBinding
 import com.neuralbit.letsnote.entities.NoteFire
 import com.neuralbit.letsnote.ui.allNotes.AllNotesViewModel
