+++ app/src/main/res/layout/activity_add_edit_note.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/coordinatorlayout"
-    tools:context=".AddEditNoteActivity">
+    tools:context=".ui.addEditNote.AddEditNoteActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
