+++ app/src/main/res/layout/activity_fingerprint.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fingerprint">
+    tools:context=".ui.addEditNote.Fingerprint">
 
     <TextView
         android:id="@+id/textView"
