+++ app/src/main/res/layout/activity_splash.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?attr/colorPrimary"
-    tools:context=".SplashActivity">
+    tools:context=".ui.signIn.SplashActivity">
 
     <ImageView
         android:id="@+id/imageView6"
