+++ app/src/main/res/layout/activity_terms_and_conditions.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#2A3439"
-    tools:context=".TermsAndConditions">
+    tools:context=".ui.signIn.TermsAndConditions">
 
     <WebView
         android:layout_width="match_parent"
